refactor(teacher): Unify student management pages
Consolidates the functionality from StudentListPage, TeacherDatabasePage, and DisabledStudentListPage into a single, unified StudentManagementPage.

This new page provides a streamlined experience for managing students, with the following features:
- A filter to view active, inactive, or all students.
- A single 'Edit' button that opens a modal with 'Basic' and 'Complete' tabs for different levels of detail.
- An integrated 'Enable/Disable' action to toggle student status directly from the list.
- A consolidated 'Add Student' flow.

The teacher dashboard has been updated to remove the redundant buttons, replacing them with a single 'Manage Students' button.

The old components have been deleted and the frontend routes have been updated to reflect these changes.
